requirements.txt
pytest == 8.4.2
selene == 2.0.0rc9
allure-pytest == 2.15.0
allure-python-commons~=2.15.0